Add files via upload
Write a program that adds salt-and-pepper noise to an image with a given
probability for both salt and pepper. Experiment with different noise levels and
visually inspect how the image quality changes.

Write a program that adds salt-and-pepper noise to an image with a specified
probability (for both salt and pepper). Analyze how this noise affects the
image's quality using metrics like Peak Signal-to-Noise Ratio (PSNR) and Mean
Squared Error (MSE).

Write a program that allows a user to define a 3 X 3 convolution kernel and
apply it to an image. The program should take user input for 9 kernel elements,
perform the convolution operation, and display the resulting image. Analyze
how different kernels (eg., identity, blur, sharpen) affect the image.

Upload the program file and report on above three questions. You consider any
color image of your photo